pyproject.toml
[project]
name = "rent-id"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "djangorestframework (>=3.15.2,<4.0.0)",
    "psycopg (>=3.2.6,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "pylint (>=3.3.6,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
